// app/apis/pantmig-api/models/ObjectPagedSearchRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PantmigService API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ObjectPagedSearchRequest
 */
export interface ObjectPagedSearchRequest {
    /**
     * 
     * @type {string}
     * @memberof ObjectPagedSearchRequest
     */
    cityExternalId?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ObjectPagedSearchRequest
     */
    page?: number;
    /**
     * 
     * @type {number}
     * @memberof ObjectPagedSearchRequest
     */
    pageSize?: number;
    /**
     * 
     * @type {boolean}
     * @memberof ObjectPagedSearchRequest
     */
    onlyActive?: boolean;
    /**
     * 
     * @type {number}
     * @memberof ObjectPagedSearchRequest
     */
    latitude?: number | null;
    /**
     * 
     * @type {number}
     * @memberof ObjectPagedSearchRequest
     */
    longitude?: number | null;
}

/**
 * Check if a given object implements the ObjectPagedSearchRequest interface.
 */
export function instanceOfObjectPagedSearchRequest(value: object): value is ObjectPagedSearchRequest {
    return true;
}

export function ObjectPagedSearchRequestFromJSON(json: any): ObjectPagedSearchRequest {
    return ObjectPagedSearchRequestFromJSONTyped(json, false);
}

export function ObjectPagedSearchRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ObjectPagedSearchRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'cityExternalId': json['cityExternalId'] == null ? undefined : json['cityExternalId'],
        'page': json['page'] == null ? undefined : json['page'],
        'pageSize': json['pageSize'] == null ? undefined : json['pageSize'],
        'onlyActive': json['onlyActive'] == null ? undefined : json['onlyActive'],
        'latitude': json['latitude'] == null ? undefined : json['latitude'],
        'longitude': json['longitude'] == null ? undefined : json['longitude'],
    };
}

export function ObjectPagedSearchRequestToJSON(json: any): ObjectPagedSearchRequest {
    return ObjectPagedSearchRequestToJSONTyped(json, false);
}

export function ObjectPagedSearchRequestToJSONTyped(value?: ObjectPagedSearchRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'cityExternalId': value['cityExternalId'],
        'page': value['page'],
        'pageSize': value['pageSize'],
        'onlyActive': value['onlyActive'],
        'latitude': value['latitude'],
        'longitude': value['longitude'],
    };
}