// app/apis/pantmig-api/models/RecycleListingItemResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PantmigService API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RecycleMaterialType } from './RecycleMaterialType';
import {
    RecycleMaterialTypeFromJSON,
    RecycleMaterialTypeFromJSONTyped,
    RecycleMaterialTypeToJSON,
    RecycleMaterialTypeToJSONTyped,
} from './RecycleMaterialType';

/**
 * 
 * @export
 * @interface RecycleListingItemResponse
 */
export interface RecycleListingItemResponse {
    /**
     * 
     * @type {number}
     * @memberof RecycleListingItemResponse
     */
    id?: number;
    /**
     * 
     * @type {RecycleMaterialType}
     * @memberof RecycleListingItemResponse
     */
    materialType?: RecycleMaterialType;
    /**
     * 
     * @type {number}
     * @memberof RecycleListingItemResponse
     */
    quantity?: number;
    /**
     * 
     * @type {string}
     * @memberof RecycleListingItemResponse
     */
    depositClass?: string | null;
    /**
     * 
     * @type {number}
     * @memberof RecycleListingItemResponse
     */
    estimatedDepositPerUnit?: number | null;
}



/**
 * Check if a given object implements the RecycleListingItemResponse interface.
 */
export function instanceOfRecycleListingItemResponse(value: object): value is RecycleListingItemResponse {
    return true;
}

export function RecycleListingItemResponseFromJSON(json: any): RecycleListingItemResponse {
    return RecycleListingItemResponseFromJSONTyped(json, false);
}

export function RecycleListingItemResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecycleListingItemResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'materialType': json['materialType'] == null ? undefined : RecycleMaterialTypeFromJSON(json['materialType']),
        'quantity': json['quantity'] == null ? undefined : json['quantity'],
        'depositClass': json['depositClass'] == null ? undefined : json['depositClass'],
        'estimatedDepositPerUnit': json['estimatedDepositPerUnit'] == null ? undefined : json['estimatedDepositPerUnit'],
    };
}

export function RecycleListingItemResponseToJSON(json: any): RecycleListingItemResponse {
    return RecycleListingItemResponseToJSONTyped(json, false);
}

export function RecycleListingItemResponseToJSONTyped(value?: RecycleListingItemResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'materialType': RecycleMaterialTypeToJSON(value['materialType']),
        'quantity': value['quantity'],
        'depositClass': value['depositClass'],
        'estimatedDepositPerUnit': value['estimatedDepositPerUnit'],
    };
}