// app/apis/pantmig-api/models/RecycleMaterialType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PantmigService API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const RecycleMaterialType = {
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3
} as const;
export type RecycleMaterialType = typeof RecycleMaterialType[keyof typeof RecycleMaterialType];


export function instanceOfRecycleMaterialType(value: any): boolean {
    for (const key in RecycleMaterialType) {
        if (Object.prototype.hasOwnProperty.call(RecycleMaterialType, key)) {
            if (RecycleMaterialType[key as keyof typeof RecycleMaterialType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function RecycleMaterialTypeFromJSON(json: any): RecycleMaterialType {
    return RecycleMaterialTypeFromJSONTyped(json, false);
}

export function RecycleMaterialTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecycleMaterialType {
    return json as RecycleMaterialType;
}

export function RecycleMaterialTypeToJSON(value?: RecycleMaterialType | null): any {
    return value as any;
}

export function RecycleMaterialTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): RecycleMaterialType {
    return value as RecycleMaterialType;
}